+++ src/twisted/internet/defer.py
@@ -1909,40 +1909,6 @@ def inlineCallbacks(
     return unwindGenerator
 
 
-async def bracket(
-    first: Callable[[], Union[_A, Awaitable[_A]]],
-    last: Callable[[], Union[_B, Awaitable[_B]]],
-    between: Callable[[], Union[_T, Awaitable[_T]]],
-) -> _T:
-    """
-    Invoke an action between two other actions.
-
-    :param first: A no-argument function that may return a Deferred.  It is
-        called first.
-
-    :param between: A no-argument function that may return a Deferred.  It is
-        called after ``first`` is done and completes before ``last`` is called.
-
-    :param last: A no-argument function that may return a Deferred.  It is
-        called last.
-
-    :return Deferred: A ``Deferred`` which fires with the result of
-        ``between``.
-    """
-    x = first()
-    if isinstance(x, Awaitable):
-        await x
-    try:
-        result = between()
-        if isinstance(result, Awaitable):
-            return await cast(Awaitable[_T], result)
-        return result
-    finally:
-        y = last()
-        if isinstance(y, Awaitable):
-            await y
-
-
 ## DeferredLock/DeferredQueue
